+++ examples/linux.rs
@@ -1,5 +1,5 @@
 use eeprom24x::{Eeprom24x, SlaveAddr};
-use embedded_hal::blocking::delay::DelayMs;
+use embedded_hal::delay::DelayNs;
 use linux_embedded_hal::{Delay, I2cdev};
 
 fn main() {
@@ -11,7 +11,7 @@ fn main() {
 
     eeprom.write_byte(memory_address, data).unwrap();
 
-    Delay.delay_ms(5u16);
+    Delay.delay_ms(5u32);
 
     let read_data = eeprom.read_byte(memory_address).unwrap();